+++ rcontest/option.go
@@ -1,20 +1,25 @@
 package rcontest
 
-import "time"
-
-// Settings contains configuration for RCON Server.
-type Settings struct {
-	Password             string
-	AuthResponseDelay    time.Duration
-	CommandResponseDelay time.Duration
-}
-
-// Option allows to inject settings to Server.
+// Option allows to inject Settings to Server.
 type Option func(s *Server)
 
-// SetSettings configuration for RCON Server.
+// SetSettings injects configuration for RCON Server.
 func SetSettings(settings Settings) Option {
 	return func(s *Server) {
-		s.settings = settings
+		s.Settings = settings
+	}
+}
+
+// SetAuthHandler injects HandlerFunc with authorisation data checking.
+func SetAuthHandler(handler HandlerFunc) Option {
+	return func(s *Server) {
+		s.SetAuthHandler(handler)
+	}
+}
+
+// SetCommandHandler injects HandlerFunc with commands processing.
+func SetCommandHandler(handler HandlerFunc) Option {
+	return func(s *Server) {
+		s.SetCommandHandler(handler)
 	}
 }